Basic distance working and gap detection
Distance measurement it working using the encoders. Altough it seems to be off by +- 1 tick and same with turning. This is too much error especialy for turns. I'm investigating alternative options. However all the code has been refacored and the functions put in a more senseable order. Coments were added and old code was removed or place in comments at the bottom of the lib file.